Real support for streaming, first step
