+++ zio-http/src/main/scala/zio/http/netty/client/ClientInboundHandler.scala
@@ -18,82 +18,78 @@ package zio.http.netty.client
 
 import zio._
 
-import zio.http.Response
-import zio.http.netty.{NettyFutureExecutor, NettyResponse, NettyRuntime}
+import zio.http.netty.{NettyBodyWriter, NettyFutureExecutor, NettyResponse, NettyRuntime}
+import zio.http.{Request, Response}
 
 import io.netty.channel.{ChannelHandlerContext, SimpleChannelInboundHandler}
-import io.netty.handler.codec.http.{FullHttpRequest, FullHttpResponse, HttpUtil}
+import io.netty.handler.codec.http._
 
 /**
  * Handles HTTP response
  */
 final class ClientInboundHandler(
-  zExec: NettyRuntime,
-  jReq: FullHttpRequest,
+  rtm: NettyRuntime,
+  req: Request,
+  jReq: HttpRequest,
   onResponse: Promise[Throwable, Response],
   onComplete: Promise[Throwable, ChannelState],
-  isWebSocket: Boolean,
   enableKeepAlive: Boolean,
 )(implicit trace: Trace)
-    extends SimpleChannelInboundHandler[FullHttpResponse](true) {
+    extends SimpleChannelInboundHandler[HttpObject](false) {
   implicit private val unsafeClass: Unsafe = Unsafe.unsafe
 
   override def channelActive(ctx: ChannelHandlerContext): Unit = {
-    if (isWebSocket) {
-      ctx.fireChannelActive()
-      ()
-    } else {
-      sendRequest(ctx)
-    }
+    sendRequest(ctx)
   }
 
   private def sendRequest(ctx: ChannelHandlerContext): Unit = {
-    ctx.writeAndFlush(jReq)
-    ()
+    jReq match {
+      case fullRequest: FullHttpRequest =>
+        ctx.writeAndFlush(fullRequest)
+      case _: HttpRequest               =>
+        ctx.write(jReq)
+        rtm.run(ctx, NettyRuntime.noopEnsuring) {
+          NettyBodyWriter.write(req.body, ctx).unit
+        }(Unsafe.unsafe, trace)
+        ctx.flush(): Unit
+    }
   }
 
-  override def channelRead0(ctx: ChannelHandlerContext, msg: FullHttpResponse): Unit = {
-    msg.touch("handlers.ClientInboundHandler-channelRead0")
-    // NOTE: The promise is made uninterruptible to be able to complete the promise in a error situation.
-    // It allows to avoid loosing the message from pipeline in case the channel pipeline is closed due to an error.
-    zExec.runUninterruptible(ctx, NettyRuntime.noopEnsuring) {
-      onResponse.succeed(NettyResponse.make(ctx, msg))
-    }(unsafeClass, trace)
+  override def channelRead0(ctx: ChannelHandlerContext, msg: HttpObject): Unit = {
+    msg match {
+      case response: HttpResponse =>
+        ctx.channel().config().setAutoRead(false)
+        rtm.runUninterruptible(ctx, NettyRuntime.noopEnsuring) {
+          onResponse
+            .succeed(
+              NettyResponse.make(
+                ctx,
+                response,
+                rtm,
+                onComplete,
+                enableKeepAlive && HttpUtil.isKeepAlive(response),
+              ),
+            )
+        }(unsafeClass, trace)
+      case content: HttpContent   =>
+        ctx.fireChannelRead(content): Unit
 
-    if (isWebSocket) {
-      ctx.fireChannelRead(msg.retain())
-      ctx.pipeline().remove(ctx.name()): Unit
+      case err => throw new IllegalStateException(s"Client unexpected message type: ${err}")
     }
-
-    val shouldKeepAlive = enableKeepAlive && HttpUtil.isKeepAlive(msg) || isWebSocket
-
-    if (!shouldKeepAlive) {
-      zExec.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(
-        NettyFutureExecutor
-          .executed(ctx.close())
-          .as(ChannelState.Invalid)
-          .exit
-          .flatMap(onComplete.done(_)),
-      )(unsafeClass, trace)
-    } else {
-      zExec.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(onComplete.succeed(ChannelState.Reusable))(
-        unsafeClass,
-        trace,
-      )
-    }
-
   }
 
   override def exceptionCaught(ctx: ChannelHandlerContext, error: Throwable): Unit = {
-    zExec.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(
+    rtm.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(
       onResponse.fail(error) *> onComplete.fail(error),
     )(unsafeClass, trace)
     releaseRequest()
   }
 
   private def releaseRequest(): Unit = {
-    if (jReq.refCnt() > 0) {
-      jReq.release(jReq.refCnt()): Unit
+    jReq match {
+      case fullRequest: FullHttpRequest =>
+        fullRequest.release(fullRequest.refCnt())
+      case _                            =>
     }
   }
 }
