+++ zio-http/src/main/scala/zio/http/netty/client/WebSocketClientInboundHandler.scala
@@ -1,0 +1,42 @@
+package zio.http.netty.client
+
+import zio.{Promise, Trace, Unsafe}
+
+import zio.http.Response
+import zio.http.netty.{NettyResponse, NettyRuntime}
+
+import io.netty.channel.{ChannelHandlerContext, SimpleChannelInboundHandler}
+import io.netty.handler.codec.http.FullHttpResponse
+
+final class WebSocketClientInboundHandler(
+  rtm: NettyRuntime,
+  onResponse: Promise[Throwable, Response],
+  onComplete: Promise[Throwable, ChannelState],
+)(implicit trace: Trace)
+    extends SimpleChannelInboundHandler[FullHttpResponse](true) {
+  implicit private val unsafeClass: Unsafe = Unsafe.unsafe
+
+  override def channelActive(ctx: ChannelHandlerContext): Unit = {
+    ctx.fireChannelActive()
+  }
+
+  override def channelRead0(ctx: ChannelHandlerContext, msg: FullHttpResponse): Unit = {
+    rtm.runUninterruptible(ctx, NettyRuntime.noopEnsuring) {
+      onResponse.succeed(NettyResponse.make(ctx, msg))
+    }(unsafeClass, trace)
+
+    ctx.fireChannelRead(msg.retain())
+    ctx.pipeline().remove(ctx.name()): Unit
+
+    rtm.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(onComplete.succeed(ChannelState.Reusable))(
+      unsafeClass,
+      trace,
+    )
+  }
+
+  override def exceptionCaught(ctx: ChannelHandlerContext, error: Throwable): Unit = {
+    rtm.runUninterruptible(ctx, NettyRuntime.noopEnsuring)(
+      onResponse.fail(error) *> onComplete.fail(error),
+    )(unsafeClass, trace)
+  }
+}
